move login helper functions to user_helpers
