+++ views/pages/login-register.ejs
@@ -7,6 +7,7 @@
 
   <%- include('../partials/_head-styles.ejs'); %>
   <%- include('../partials/_head-scripts.ejs'); %>
+  <script defer src="/scripts/login-register.js"></script>
 
   <title>QUOKKA? - Home</title>
 </head>
